+++ src/Framework/tests/Routing/Commands/RouteListCommandTest.php
@@ -1,0 +1,26 @@
+<?php
+
+/**
+ * Aphiria
+ *
+ * @link      https://www.aphiria.com
+ * @copyright Copyright (C) 2022 David Young
+ * @license   https://github.com/aphiria/aphiria/blob/1.x/LICENSE.md
+ */
+
+declare(strict_types=1);
+
+namespace Aphiria\Framework\Tests\Routing\Commands;
+
+use Aphiria\Framework\Routing\Commands\RouteListCommand;
+use PHPUnit\Framework\TestCase;
+
+class RouteListCommandTest extends TestCase
+{
+    public function testCorrectValuesAreSetInConstructor(): void
+    {
+        $command = new RouteListCommand();
+        $this->assertSame('route:list', $command->name);
+        $this->assertSame('Lists the routes in your app', $command->description);
+    }
+}